+++ output/kotlin/models/interface/LogEntry.kt
@@ -2,7 +2,7 @@ package `interface`
 
 interface LogEntry {
     val deviceState: DeviceState
-    val timeFired: Int
+    val timeFired: Long
     val values: Map<String, String>
     val tag: String
     val stackTrace: String
