+++ output/kotlin/models/serializable_data_class/LogEntry.kt
@@ -3,7 +3,7 @@ package `serializable_data_class`
 @Serializable
 data class LogEntry(
     @SerialName("device_state") val deviceState: DeviceState,
-    @SerialName("time_fired") val timeFired: Int,
+    @SerialName("time_fired") val timeFired: Long,
     @SerialName("values") val values: Map<String, String>,
     @SerialName("tag") val tag: String,
     @SerialName("stack_trace") val stackTrace: String,
